Tests/LibWeb/Text/input/input-method-insert-text.html

<!DOCTYPE html>
<script src="include.js"></script>
<input id="input" value="abc">
<textarea id="textarea">xyz</textarea>
<div id="editable" contenteditable="true">qwe</div>
<input id="readonly_input" value="ro" readonly>
<input id="preedit_input" value="">
<script>
    asyncTest(async done => {
        // Spec for <input>/<textarea> fires 'input' via a queued element task. Wait for it.
        const waitForEvent = (target, type) => new Promise(resolve => {
            target.addEventListener(type, e => resolve({ inputType: e.inputType, data: e.data }), { once: true });
        });

        // <input>: committed IME text is inserted via insertText and eventually fires an input event.
        const input = document.getElementById("input");
        input.focus();
        input.setSelectionRange(input.value.length, input.value.length);
        let event_promise = waitForEvent(input, "input");
        internals.insertTextFromInputMethod("def");
        let event = await event_promise;
        println(`input.value after IME insert: ${JSON.stringify(input.value)}`);
        println(`input event on <input>: ${JSON.stringify(event)}`);

        // <textarea>: same path.
        const textarea = document.getElementById("textarea");
        textarea.focus();
        textarea.setSelectionRange(textarea.value.length, textarea.value.length);
        event_promise = waitForEvent(textarea, "input");
        internals.insertTextFromInputMethod("yyy");
        event = await event_promise;
        println(`textarea.value after IME insert: ${JSON.stringify(textarea.value)}`);
        println(`input event on <textarea>: ${JSON.stringify(event)}`);

        // contenteditable: fires input synchronously through the editing pipeline.
        const editable = document.getElementById("editable");
        editable.focus();
        getSelection().selectAllChildren(editable);
        getSelection().collapseToEnd();
        const collected_editable_events = [];
        editable.addEventListener("input", e => collected_editable_events.push({ inputType: e.inputType, data: e.data }), { once: true });
        internals.insertTextFromInputMethod("rty");
        println(`editable.textContent after IME insert: ${JSON.stringify(editable.textContent)}`);
        println(`input event on contenteditable: ${JSON.stringify(collected_editable_events[0])}`);

        // No focus on any editable: must be a no-op (no exception, no event).
        document.activeElement?.blur();
        let no_focus_threw = false;
        try {
            internals.insertTextFromInputMethod("ignored");
        } catch (e) {
            no_focus_threw = true;
        }
        println(`unfocused insert threw: ${no_focus_threw}`);

        // <input readonly>: not editable, IME insert must not change the value.
        const ro = document.getElementById("readonly_input");
        ro.focus();
        internals.insertTextFromInputMethod("nope");
        println(`readonly value unchanged: ${ro.value === "ro"}`);

        // currentCaretRect returns a sensible rect on a focused editable, null otherwise.
        input.focus();
        input.setSelectionRange(0, 0);
        const rect = internals.currentCaretRect();
        println(`focused-input caret rect non-null: ${rect !== null}`);
        println(`focused-input caret rect has positive height: ${rect && rect.height > 0}`);

        document.activeElement?.blur();
        const no_focus_rect = internals.currentCaretRect();
        println(`unfocused caret rect: ${no_focus_rect === null ? "null" : "non-null"}`);

        // Empty editable: the caret rect must still be sensible.
        const empty = document.getElementById("preedit_input"); // value is ""
        empty.focus();
        empty.setSelectionRange(0, 0);
        const empty_rect = internals.currentCaretRect();
        println(`empty-input caret rect non-null: ${empty_rect !== null}`);
        println(`empty-input caret rect has positive height: ${empty_rect && empty_rect.height > 0}`);

        // replaceBeforeCaretInCodeUnits — the per-keystroke preedit-swap path used by IME composition.
        // Simulate a Pinyin sequence: type "n", then "i" (extending preedit), then commit "你" (replacing
        // the full "ni" preedit), then commit "好" (no preedit outstanding).
        const preedit = document.getElementById("preedit_input");
        preedit.focus();
        preedit.setSelectionRange(0, 0);

        // First keystroke: insert "n" with 0 chars to replace.
        internals.insertTextFromInputMethod("n", 0);
        println(`after first preedit keystroke: ${JSON.stringify(preedit.value)}`);

        // Second keystroke: extend preedit — replace the previous "n" with "ni".
        internals.insertTextFromInputMethod("ni", 1);
        println(`after extending preedit: ${JSON.stringify(preedit.value)}`);

        // Commit: replace the "ni" preedit (2 chars) with the selected candidate "你".
        internals.insertTextFromInputMethod("你", 2);
        println(`after first commit: ${JSON.stringify(preedit.value)}`);

        // Second commit immediately after, no preedit outstanding.
        internals.insertTextFromInputMethod("好", 0);
        println(`after second commit: ${JSON.stringify(preedit.value)}`);

        // unmarkText path: replace marked text with empty string (IME abort).
        internals.insertTextFromInputMethod("abc", 0);
        internals.insertTextFromInputMethod("ab", 3); // 3-char preedit → replaced with 2-char preedit
        internals.insertTextFromInputMethod("", 2);   // abort: clear the 2-char preedit
        println(`after abort sequence: ${JSON.stringify(preedit.value)}`);

        // Edge case: replace count larger than what's there — must not underflow.
        preedit.value = "ab";
        preedit.setSelectionRange(2, 2);
        internals.insertTextFromInputMethod("Z", 10);
        println(`after over-replace: ${JSON.stringify(preedit.value)}`);

        done();
    });
</script>
